Fixup ping source dataMaxSize configmap (#4894)
* Fixup ping source dataMaxSize configmap

- Configmap should be basic key/value pair, not a string for a yaml struct
- Provide example dataMaxSize configmap config
- Correct example-checksum annotation
- No mashalling from yaml needed, strconv directly to int64
- Add unit testing for ConfigMap reading
- Remove erroneous broker parsing comments/function from copied implementation
- Rename configmap to match naming conventions
- update bsize len assignment to match int64 size

* Update dataMaxSize example to -1 (default value)

* nit:  remove  extra  space  in  pingsource  configmap  example

* rename PingDataMaxSize DefaultDataMaxSize

+++ pkg/apis/sources/v1beta2/ping_validation.go
@@ -63,7 +63,7 @@ func (cs *PingSourceSpec) Validate(ctx context.Context) *apis.FieldError {
 	if cs.Data != "" && cs.DataBase64 != "" {
 		errs = errs.Also(apis.ErrMultipleOneOf("data", "dataBase64"))
 	} else if cs.DataBase64 != "" {
-		if bsize := len(cs.DataBase64); pingDefaults.DataMaxSize > -1 && bsize > pingDefaults.DataMaxSize {
+		if bsize := int64(len(cs.DataBase64)); pingDefaults.DataMaxSize > -1 && bsize > pingDefaults.DataMaxSize {
 			fe := apis.ErrInvalidValue(fmt.Sprintf("the dataBase64 length of %d bytes exceeds limit set at %d.", bsize, pingDefaults.DataMaxSize), "dataBase64")
 			errs = errs.Also(fe)
 		}
@@ -80,7 +80,7 @@ func (cs *PingSourceSpec) Validate(ctx context.Context) *apis.FieldError {
 			}
 		}
 	} else if cs.Data != "" {
-		if bsize := len(cs.Data); pingDefaults.DataMaxSize > -1 && bsize > pingDefaults.DataMaxSize {
+		if bsize := int64(len(cs.Data)); pingDefaults.DataMaxSize > -1 && bsize > pingDefaults.DataMaxSize {
 			fe := apis.ErrInvalidValue(fmt.Sprintf("the data length of %d bytes exceeds limit set at %d.", bsize, pingDefaults.DataMaxSize), "data")
 			errs = errs.Also(fe)
 		}
